[Fix] Fetch all Satellites available on N2YO (#44)
Previously, only satellites that are right now in space are represented
in the chart. That makes the numbers for years prior to 2024 wrong. The
problem originate as N2YO only shows those satellites for a
constellation that are currently trackable. Untrackable satellites are
not included. To fix this problem, one needs to go over all satellite
IDs available on N2YO and put them to the proper constellation.