[Thesis] Motivation and Outline of thesis (#240)
This pull request includes several significant updates to the thesis
document, primarily focusing on enhancing the introduction and
structuring the thesis content. The most important changes include the
addition of new sections, content, and references to improve the overall
clarity and depth of the thesis.

### Content Additions and Enhancements:

*
[`tex/thesis/chapters/1-introduction/motivation.tex`](diffhunk://#diff-befa1ae0d3441bcc2ff7d93e41bd08175028c2efcb1de8cb349d4e684ea4b3c7R2-R45):
Added a detailed motivation section discussing the role of the internet,
challenges with current infrastructure, and the potential of satellite
communication.
*
[`tex/thesis/chapters/1-introduction/thesis-outline.tex`](diffhunk://#diff-1bd70ef68b034b51c93477e871736e6de06d0de1bb16edd9273c053499362223R1-R29):
Introduced a new section outlining the structure of the thesis,
including chapters on introduction, background, methodology, results,
and conclusion.

### Structural Changes:

*
[`tex/thesis/content.tex`](diffhunk://#diff-9013ecc898f9ca9fb2d9217a507bb9c58ebc8615f5f80d7c377bef4a83e1e13fL6-R9):
Updated the `Introduction` chapter to include the motivation and thesis
outline sections. Added labels to chapters for better referencing.
[[1]](diffhunk://#diff-9013ecc898f9ca9fb2d9217a507bb9c58ebc8615f5f80d7c377bef4a83e1e13fL6-R9)
[[2]](diffhunk://#diff-9013ecc898f9ca9fb2d9217a507bb9c58ebc8615f5f80d7c377bef4a83e1e13fL18-R18)
[[3]](diffhunk://#diff-9013ecc898f9ca9fb2d9217a507bb9c58ebc8615f5f80d7c377bef4a83e1e13fL35-R35)
[[4]](diffhunk://#diff-9013ecc898f9ca9fb2d9217a507bb9c58ebc8615f5f80d7c377bef4a83e1e13fL93-R93)

### Reference Updates:

*
[`tex/thesis/references/references.bib`](diffhunk://#diff-c73e1eef5b86ab73ff987022a52036ea059c9d66172c4c02ae4a8850bd2ef1b6R675-R687):
Added a new reference for a 1980 article by Kenneth Davies on satellite
radio beacon studies.