[Thesis] Removing OONI chapter and adding referenced table (#243)
This pull request includes several changes to the thesis document,
primarily focusing on the removal of a data source and the addition of a
new table in the appendix.

### Removal of Data Source:
*
[`tex/thesis/chapters/3-methodology/measurement-platforms.tex`](diffhunk://#diff-81792dd1a6770d1816572ed421f4c842f367168c68c730a155d4eb54605abf59L93-L95):
Removed the section on OONI, indicating it is no longer considered
necessary for the thesis.
*
[`tex/thesis/chapters/3-methodology/methodology.tex`](diffhunk://#diff-a16fb88c0e3db5b88c3dd493c5ec537250d54010e5f5c1c846c9d075b54b7390L8-R8):
Updated the data collection section to reflect the removal of OONI as a
data source.

### Addition of New Table:
*
[`tex/thesis/chapters/appendix/figures.tex`](diffhunk://#diff-53a85715049ac59ea37c8ade9049e38ebacf1f52efb9c1a600101613a2063798R1-R36):
Added a new table showing the number of probes per country with ASN
14593 on RIPE Atlas.